Abscence1

+++ Data/employe.cs
@@ -28,7 +28,7 @@ namespace GestionAbscences.Data
         public Nullable<System.DateTime> DateFin { get; set; }
         public string password { get; set; }
         public Nullable<int> nbjours { get; set; }
-        public Nullable<double> nbjoursR { get; set; }
+        public Nullable<int> nbjoursR { get; set; }
         public string matricule { get; set; }
         public string affectation { get; set; }
         public string role { get; set; }
